requirements.txt
flask
openai-whisper
torch
numpy
ffmpeg-python
